requirements.txt
# Core dependencies
streamlit>=1.30.0
langchain>=0.1.0
openai>=1.10.0
chromadb>=0.4.22

# Document processing
python-pptx>=0.6.21
python-docx>=0.8.11
PyPDF2>=3.0.0
Pillow>=10.0.0
markitdown>=0.1.0  # For PDF processing

# Vector store and embeddings
numpy>=1.24.0
pandas>=2.0.0

# Development tools
python-dotenv>=1.0.0
pytest>=7.4.0  # For testing
black>=23.0.0  # For code formatting
isort>=5.12.0  # For import sorting
mypy>=1.5.0   # For type checking

# Web search
smol-dev-js>=0.1.0  # For web search functionality 